# docs/swagger.yaml
definitions:
  dto.GoogleLoginDTO:
    properties:
      idToken:
        type: string
    required:
    - idToken
    type: object
info:
  contact: {}
  description: Made By Walter.Fish
  title: Tabinote API
  version: 0.0.1
paths:
  /api/v1/auth/google:
    post:
      parameters:
      - description: payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.GoogleLoginDTO'
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "401":
          description: 失敗
          schema:
            type: string
      summary: google登入
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      description: get my userInfo
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "401":
          description: 失敗
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Me
      tags:
      - Auth
  /api/v1/health/db:
    get:
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "401":
          description: 失敗
          schema:
            type: string
      summary: 測試
      tags:
      - Health 伺服器狀態檢查
  /api/v1/health/parseToken:
    post:
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "401":
          description: 失敗
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 測試token解碼
      tags:
      - Health 伺服器狀態檢查
  /api/v1/health/uuid:
    get:
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "401":
          description: 失敗
          schema:
            type: string
      summary: 生成一個uuid
      tags:
      - Health 伺服器狀態檢查
  /api/v1/user/login:
    post:
      description: 用戶登入詳情描述
      responses:
        "200":
          description: 登入成功
          schema:
            type: string
        "401":
          description: 登入失敗
          schema:
            type: string
      summary: 用戶登入
      tags:
      - 用戶管理
securityDefinitions:
  BearerAuth:
    description: 請在輸入框輸入 "Bearer {你的Token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"